Business search getting a response

+++ app/src/main/java/com/example/vetted/modells/Businesses.java
@@ -6,13 +6,13 @@ import java.util.List;
 public class Businesses{
   @SerializedName("distance")
   @Expose
-  private Integer distance;
+  private Double distance;
   @SerializedName("image_url")
   @Expose
   private String image_url;
   @SerializedName("rating")
   @Expose
-  private Integer rating;
+  private Double rating;
   @SerializedName("coordinates")
   @Expose
   private Coordinates coordinates;
@@ -52,10 +52,10 @@ public class Businesses{
   public Businesses(){
   }
 
-  public void setDistance(Integer distance){
+  public void setDistance(Double distance){
    this.distance=distance;
   }
-  public Integer getDistance(){
+  public Double getDistance(){
    return distance;
   }
   public void setImage_url(String image_url){
@@ -64,10 +64,10 @@ public class Businesses{
   public String getImage_url(){
    return image_url;
   }
-  public void setRating(Integer rating){
+  public void setRating(Double rating){
    this.rating=rating;
   }
-  public Integer getRating(){
+  public Double getRating(){
    return rating;
   }
   public void setCoordinates(Coordinates coordinates){
